fear(portal): save role list
