+++ vortex-app/vortex-app-portal/src/components/Access/Authenticate.tsx
@@ -7,7 +7,7 @@ import { filter, get } from 'lodash'
 import { useGetUserAuthDetail, useGetUserRole } from '@/hooks/user'
 import { ENV } from '@/constant'
 import { useAppStore } from '@/stores/app.store'
-import type { AuthUser } from '@/stores/app.store'
+import type { AuthUser } from '@/stores/type'
 
 interface AuthenticateProps {
   children: ReactNode
@@ -15,7 +15,7 @@ interface AuthenticateProps {
 
 const Authenticate = ({ children }: AuthenticateProps) => {
   const { isLoading, isAuthenticated, user } = useAuth0()
-  const { currentAuth0User, setCurrentAuth0User, setUser } = useAppStore()
+  const { currentAuth0User, setCurrentAuth0User, setUser, setRoleList } = useAppStore()
   const navigate = useNavigate()
 
   const { data: userData } = useGetUserAuthDetail()
@@ -26,10 +26,9 @@ const Authenticate = ({ children }: AuthenticateProps) => {
     const roleList = roleData?.data
     if (!userDetail) return
     const companyId = get(userDetail, 'companies[0].id', '')
-
     const roleIds = get(userDetail, ['linkUserCompany', companyId, 'roleIds'], [])
     const accessRole = filter(roleList, (r) => roleIds.includes(r.id) || r.systemDefault)
-
+    setRoleList(roleList)
     setUser({ ...userDetail, accessRole })
   }, [userData, roleData])
